requirements.txt
pyserial==3.5
websockets==11.0.3
